Numerous fixes

+++ client/include/smtp-sockets.h
@@ -1,9 +1,12 @@
 #ifndef __SMTP_SOCKETS_H__
 #define __SMTP_SOCKETS_H__
 
+#include <stdbool.h>
+#include <sys/socket.h>
+
 void socket_set_nonblocking(int socket, bool is_nonblocking);
 int create_local_socket_pair(int *out_sockets);
-int socket_write(int socket, char *buf, int len);
-int socket_read(int socket, char *buf, int len);
+ssize_t socket_write(int socket, char *buf, int len);
+ssize_t socket_read(int socket, char *buf, int len);
 
 #endif
